+++ runtime/lib/schedule_microtask_patch.dart
@@ -24,3 +24,7 @@ class _ScheduleImmediate {
 void _setScheduleImmediateClosure(_ScheduleImmediateClosure closure) {
   _ScheduleImmediate._closure = closure;
 }
+
+void _ensureScheduleImmediate() {
+  _AsyncRun._scheduleImmediate(_startMicrotaskLoop);
+}